fix(lint): resolve linter errors
Prompt: fix linter errors

- Removed unused imports of areDiceRolling
- Refactored GameContext to improve fast refresh:
  - Separated context definition into GameContext.tsx
  - Moved provider implementation to GameContextProvider.tsx
  - Updated imports throughout the codebase
  - Simplified useGameState hook
