+++ src/components/game/GameBoard/GameBoard.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import type { GameState, ScoreCategory } from '../../../types/game';
 import { Die } from '../../dice/Die/Die';
-import { areDiceRolling } from '../../../utils/diceUtils';
-
 import { MultiPlayerScorecard } from '../../scorecard/MultiPlayerScorecard/MultiPlayerScorecard';
 import { calculateScore, canScoreCategory } from '../../../utils/scoring';
 import './GameBoard.css';
