+++ src/hooks/useGameState.ts
@@ -1,11 +1,6 @@
-import { useContext } from 'react';
-import { GameContext } from '../contexts/GameContext';
+import { useGameContext } from '../contexts/GameContext';
 
 // Custom hook to use game context
 export const useGameState = () => {
-  const context = useContext(GameContext);
-  if (!context) {
-    throw new Error('useGameState must be used within a GameProvider');
-  }
-  return context;
+  return useGameContext();
 };
